Added the Assembly Resolver to replace hard coded assembly paths.
Before all compiled assmblies only got references to UnityEngine, UnityEditor, and System when compiling. Now we have the assembly resolver which tries to find the  assemblies in the current app domine. Current does not take into account dependices inside of none generated assemblies. 
Added a the starting of a helpful attribute to find assemblies in the current domain.